lkjlkj

+++ production/solver/cuda_async/src/ncg.cu
@@ -17,13 +17,13 @@ using namespace std;
 
 int main(int argc, char* argv[]) {
 
-	if (argc > 1) { _GLB_N_ = _GLB_N_ * (double)atoi(argv[1]); }
+	if (argc > 1) { _GLB_N_ = (long int) _GLB_N_ * atof(argv[1]); }
 
-	if (argc > 2) { _GLB_ITR_ = _GLB_ITR_ * (double)atoi(argv[2]); }
+	if (argc > 2) { _GLB_ITR_ = (long int)_GLB_ITR_ * atof(argv[2]); }
 
-	if (argc > 3) { _GLB_ITR_LINE_ = _GLB_ITR_LINE_ * (double)atoi(argv[3]); }
+	if (argc > 3) { _GLB_ITR_LINE_ = (long int)_GLB_ITR_LINE_ * atof(argv[3]); }
 
-	if (argc > 4) { _GLB_EPS_ = _GLB_EPS_ * (double)atoi(argv[4]); }
+	if (argc > 4) { _GLB_EPS_ = (double) _GLB_EPS_ * atof(argv[4]); }
 
 	cout << _GLB_ITR_ << endl;
